requirements.txt
anthropic==0.40.0
openai>=1.0.0
python-dotenv==1.0.0
scipy>=1.11.0
tiktoken==0.5.0
click>=8.1.0
rich>=13.7.0